fix isDefaultBranch + fix delete product

+++ src/mpModules/marketConfig/marketConfigService.js
@@ -723,6 +723,18 @@ module.exports.deleteProductService = async (result) => {
             message: `Không tồn tại sản phẩm trên chợ có id = ${id}`
         }
     }
+    const findProductMarketOrder = await models.MarketOrderProduct.findOne({
+        where:{
+            marketProductId: id,
+        }
+    });
+    if(findProductMarketOrder){
+        return {
+            error: true,
+            code: HttpStatusCode.BAD_REQUEST,
+            message: `Sản phẩm đã được đặt hàng, không thể xóa`
+        }
+    }
     const t = await models.sequelize.transaction(async (t) => {
         await models.MarketProductBatch.destroy({
             where: {
